Add failing test for Delete with a where clause evaluating to false
The Delete method fails with a `System.Data.SqlClient.SqlException` when the where clause evaluates to false. Obviously, it would be better not to call the Delete method at all if the expression is known to evaluate to false but I think it should not fail nonetheless.

I would love to propose a solution to this issue but unfortunately, I'm not yet familiar enough with LINQ to Entities.