Fix a bug with formatting nullable types
The `FormatValue()` method can be called multiple times on the same entity (at least once for formatting `OldValue` and once for formatting `NewValue`). If it fails to find an appropriate formatter the first time (which can happen if the value is null, for example), it adds the (null) formatter to the dictionary. Then, when it's called the next time, `TryGetValue` finds a formatter in the dictionary (but it's null) causing it to revert to the default formatting even when it shouldn't.

Fixed by ensuring the formatter isn't null before adding it to the dictionary.

The problem described above can easily be reproduced by attempting to format any nullable type a first time when the value is null and a second time when it isn't. For example:

```
AuditManager.DefaultConfiguration.FormatType<DateTime?>(d => d?.ToString("yyyy-MM-dd"));
```